Update README.md
Hangman is a classic word-guessing game. The objective is to guess the hidden word by suggesting letters one at a time. It’s both fun and educational, commonly used to improve vocabulary and logical thinking.
.

🖥️ 1. Display Size in Terminal is 
Most displays easily fit in an 80 x 24 terminal window  80 characters wide, 24 lines tall The hangman drawing takes around

Width: ~13 characters

Height: ~9 lines
The game randomly selects a word from a list. For example: "python".

Display Setup:
The word is hidden using underscores:

The player guesses a letter 


If the letter is in the word, it replaces the underscore:


If the guess is wrong, a part of a hangman figure is drawn (like head, body, arms, legs).


Players get limited attempts (usually 6-7). The game ends if:

The full word is guessed (Win 🏆)

The drawing completes (Lose ❌)

it Covers key programming concepts:

Loops and conditionals

Strings and lists

Random module

User input


In the future, this game can be improved by adding features like:

A graphical interface using Tkinter or Pygame

A larger word database or category selection

Sound effects and animations

This project was both fun and educational, and it’s a great stepping stone into game development and Python programming.